# structure-service/internal/infrastructure/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for managing university structure hierarchy",
        "title": "Structure Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/departments/managers": {
            "get": {
                "description": "Возвращает список всех назначений операторов на подразделения",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department-managers"
                ],
                "summary": "Получить все назначения операторов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DepartmentManager"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Назначает оператора на филиал или факультет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department-managers"
                ],
                "summary": "Назначить оператора на подразделение",
                "parameters": [
                    {
                        "description": "Данные назначения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AssignOperatorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.DepartmentManager"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/departments/managers/{id}": {
            "delete": {
                "description": "Удаляет назначение оператора на подразделение",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department-managers"
                ],
                "summary": "Удалить назначение оператора",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID назначения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/groups/{group_id}/chat": {
            "put": {
                "description": "Обновляет chat_id для группы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Связать группу с чатом",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ID чата",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.LinkGroupToChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/import/excel": {
            "post": {
                "description": "Импортирует структуру вуза из Excel файла",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Импортировать структуру из Excel",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Excel файл со структурой",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ImportResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/structure": {
            "post": {
                "description": "Создает или находит все элементы структуры (университет, филиал, факультет, группа)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "structure"
                ],
                "summary": "Создать полную структуру",
                "parameters": [
                    {
                        "description": "Данные структуры",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateStructureRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateStructureResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/universities": {
            "get": {
                "description": "Возвращает список всех вузов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Получить все вузы",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.University"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новый вуз",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Создать вуз",
                "parameters": [
                    {
                        "description": "Данные вуза",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.University"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.University"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/universities/{id}": {
            "get": {
                "description": "Возвращает информацию о вузе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Получить вуз по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вуза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.University"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/universities/{university_id}/structure": {
            "get": {
                "description": "Возвращает иерархическую структуру вуза (университет -\u003e филиал -\u003e факультет -\u003e группа -\u003e чат)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "structure"
                ],
                "summary": "Получить структуру вуза",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вуза",
                        "name": "university_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StructureNode"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Chat": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "max_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "domain.DepartmentManager": {
            "type": "object",
            "properties": {
                "assigned_at": {
                    "type": "string"
                },
                "assigned_by": {
                    "description": "User ID куратора",
                    "type": "integer"
                },
                "branch_id": {
                    "type": "integer"
                },
                "employee_id": {
                    "type": "integer"
                },
                "faculty_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "domain.ImportResult": {
            "type": "object",
            "properties": {
                "created": {
                    "description": "Количество созданных записей",
                    "type": "integer"
                },
                "errors": {
                    "description": "Список ошибок",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed": {
                    "description": "Количество неудачных записей",
                    "type": "integer"
                },
                "updated": {
                    "description": "Количество обновленных записей",
                    "type": "integer"
                }
            }
        },
        "domain.StructureNode": {
            "type": "object",
            "properties": {
                "chat": {
                    "$ref": "#/definitions/domain.Chat"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.StructureNode"
                    }
                },
                "course": {
                    "type": "integer"
                },
                "group_num": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "description": "\"university\", \"branch\", \"faculty\", \"group\"",
                    "type": "string"
                }
            }
        },
        "domain.University": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "foiv": {
                    "description": "ФОИВ",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inn": {
                    "type": "string"
                },
                "kpp": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "http.AssignOperatorRequest": {
            "type": "object",
            "properties": {
                "assigned_by": {
                    "type": "integer"
                },
                "branch_id": {
                    "type": "integer"
                },
                "employee_id": {
                    "type": "integer"
                },
                "faculty_id": {
                    "type": "integer"
                }
            }
        },
        "http.LinkGroupToChatRequest": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "integer"
                }
            }
        },
        "usecase.CreateStructureRequest": {
            "type": "object",
            "properties": {
                "branch_name": {
                    "type": "string"
                },
                "course": {
                    "type": "integer"
                },
                "faculty_name": {
                    "type": "string"
                },
                "foiv": {
                    "type": "string"
                },
                "group_number": {
                    "type": "string"
                },
                "inn": {
                    "type": "string"
                },
                "kpp": {
                    "type": "string"
                },
                "org_name": {
                    "type": "string"
                }
            }
        },
        "usecase.CreateStructureResponse": {
            "type": "object",
            "properties": {
                "branch_id": {
                    "type": "integer"
                },
                "faculty_id": {
                    "type": "integer"
                },
                "group_id": {
                    "type": "integer"
                },
                "university_id": {
                    "type": "integer"
                }
            }
        }
    }
}